switch all OS versions to the native badge
