--- WooCommerce/Classes/ViewRelated/TabBar/WooTab+Tag.swift
@@ -1,21 +1,0 @@
-import Foundation
-
-extension WooTab {
-    /// An integer identifier for each tab
-    var identifierNumber: Int {
-        switch self {
-        case .myStore:
-            return 0
-        case .orders:
-            return 1
-        case .products:
-            return 2
-        case .bookings:
-            return 3
-        case .pointOfSale:
-            return 4
-        case .hubMenu:
-            return 5
-        }
-    }
-}
